--- lib/badge/badge.interface.ts
@@ -1,6 +1,0 @@
-export interface Props {
-  modifiers?: string;
-  className?: string;
-  style?: React.CSSProperties;
-  value?: string | number;
-}
