--- lib/button/button.interface.ts
@@ -1,17 +1,0 @@
-import { TIcon } from '../icon/icon.interface';
-
-type TType = 'button' | 'submit' | 'reset';
-
-export interface Props {
-  id?: string;
-  action: string;
-  href?: string;
-  type?: TType;
-  name?: string;
-  modifiers?: string;
-  className?: string;
-  style?: React.CSSProperties;
-  disabled?: boolean;
-  icon?: TIcon;
-  onClick?: () => void;
-}
