--- lib/card/card.interface.ts
@@ -1,8 +1,0 @@
-export interface Props {
-  children: React.ReactNode;
-  className?: string;
-  href?: string;
-  modifiers?: string;
-  style?: React.CSSProperties;
-  tabIndex?: boolean;
-}
