--- lib/divider/divider.interface.ts
@@ -1,5 +1,0 @@
-export interface Props {
-  className?: string;
-  modifiers?: string;
-  style?: React.CSSProperties;
-}
