--- lib/grid/grid.interface.ts
@@ -1,10 +1,0 @@
-type TTag = 'div' | 'section';
-
-export interface Props {
-  children: React.ReactNode;
-  className?: string;
-  maxWidth?: boolean;
-  modifiers?: string;
-  style?: React.CSSProperties;
-  tag?: TTag;
-}
