--- lib/preloader/preloader.interface.ts
@@ -1,5 +1,0 @@
-export interface Props {
-  active?: boolean;
-  className?: string;
-  style?: React.CSSProperties;
-}
