--- lib/table-body/table-body.interface.ts
@@ -1,5 +1,0 @@
-export interface Props {
-  children: React.ReactNode;
-  className?: string;
-  style?: React.CSSProperties;
-}
