+++ lib/table-head/table-head.tsx
@@ -1,8 +1,12 @@
 import * as React from 'react';
 
-import { Props } from './table-head.interface';
-
 import BEM from '../utilities/ts/bem';
+
+interface Props {
+  children: React.ReactNode;
+  className?: string;
+  style?: React.CSSProperties;
+}
 
 const TableHead: React.FC<Props> = props => {
   const bem = BEM('thead');
