+++ lib/table-row/table-row.tsx
@@ -1,8 +1,13 @@
 import * as React from 'react';
 
-import { Props } from './table-row.interface';
-
 import BEM from '../utilities/ts/bem';
+
+interface Props {
+  children: React.ReactNode;
+  className?: string;
+  modifiers?: string;
+  style?: React.CSSProperties;
+}
 
 const TableRow: React.FC<Props> = props => {
   const bem = BEM('tr');
