--- lib/table/table.interface.ts
@@ -1,6 +1,0 @@
-export interface Props {
-  children: React.ReactNode;
-  className?: string;
-  modifiers?: string;
-  style?: React.CSSProperties;
-}
