--- lib/uniform/uniform.interface.ts
@@ -1,9 +1,0 @@
-type TTag = 'div' | 'main' | 'section';
-
-export interface Props {
-  children: React.ReactNode;
-  className?: string;
-  mode?: string;
-  style?: React.CSSProperties;
-  tag?: TTag;
-}
